+++ CHANGELOG.md
@@ -40,6 +40,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   **prepended** to the default system `PATH` to avoid conflicts with installed
   distro binaries which have the same name as aerc builtin filters (e.g.
   `/usr/bin/colorize`).
+- `:export-mbox` only exports marked messages, if any. Otherwise it exports
+  everything, as usual.
 
 ## [0.15.2](https://git.sr.ht/~rjarry/aerc/refs/0.15.2) - 2023-05-11